requirements.txt
PyMuPDF
python-pptx
paddleocr
paddlepaddle
camelot-py[cv]
opencv-python-headless
requests
Pillow

# Structuring module
openai
pydantic
google-genai
